+++ js/util.js
@@ -1,3 +1,5 @@
+const body = document.querySelector('body');
+
 function getRandomInteger (min, max) {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -19,4 +21,4 @@ function createIdGenerator () {
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-export {getRandomInteger, getRandomArrayElement, createIdGenerator, isEscapeKey};
+export {getRandomInteger, getRandomArrayElement, createIdGenerator, isEscapeKey, body};
